refactor(app-layout): :truck: centralize layout structure and migrate UI shell components
Move `Header`, `Sidebar`, and `Notification` into a shared `AppLayout` to enforce a consistent authenticated UI structure. Simplify routing by wrapping protected pages in a single layout and removing duplicated visual scaffolding from individual pages. Update route structure to reflect new layout hierarchy.
